Decouple runners and model variants
