+++ models/demos/deepseek_v3_d_p/tests/test_mla.py
@@ -17,7 +17,8 @@ from ttnn.device import is_blackhole
 
 import ttnn
 from models.demos.deepseek_v3_d_p.reference.mla_reference import create_mla_reference
-from models.demos.deepseek_v3_d_p.tests.model_variants import MODEL_VARIANTS, run_reference_mla
+from models.demos.deepseek_v3_d_p.tests.model_variants import MODEL_VARIANTS
+from models.demos.deepseek_v3_d_p.tests.reference_runners import run_reference_mla
 from models.demos.deepseek_v3_d_p.tt.mla import ttMLA
 from models.demos.deepseek_v3_d_p.tt.mla.rope import RotarySetup
 from models.demos.deepseek_v3_d_p.tt.mla.utils import (
